Moved all of standards.py to data.py
